+++ src/shared/messages.h
@@ -213,7 +213,13 @@ enum {
      * Send ramwatch string to display on OSD
      * Argument: size_t (string length) then char[len]
      */
-    MSGN_RAMWATCH
+    MSGN_RAMWATCH,
+
+    /*
+     * Send coords of new mouse location
+     * Arguments: two ints 
+     */
+    MSGN_CALIBRATE_MOUSE,
 };
 
 #endif